Feature/87273

See merge request edusoho/edusoho!13646

+++ src/Biz/Course/Service/Impl/CourseDeleteServiceImpl.php
@@ -90,9 +90,11 @@ class CourseDeleteServiceImpl extends BaseService implements CourseDeleteService
     {
         $this->beginTransaction();
         try {
-            $this->getProductMallGoodsRelationService()->checkMallGoods($courseId, 'course');
+            if ($this->getProductMallGoodsRelationService()->checkMallGoods([$courseId], 'course') == 'error') {
+                throw $this->createServiceException('该产品已在营销商城中上架售卖，请将对应商品下架后再进行删除操作');
+            }
 
-            $this->dispatchEvent('course.delete', new Event(['id'=>$courseId]));
+            $this->dispatchEvent('course.delete', new Event(['id' => $courseId]));
 
             $this->deleteCourseMaterial($courseId);
 
@@ -105,7 +107,7 @@ class CourseDeleteServiceImpl extends BaseService implements CourseDeleteService
             $this->getCourseDao()->delete($courseId);
 
             $this->getSchedulerService()->register([
-                'name' => 'delete_course_job'.$courseId,
+                'name' => 'delete_course_job' . $courseId,
                 'source' => SystemCrontabInitializer::SOURCE_SYSTEM,
                 'expression' => intval(time()),
                 'misfire_policy' => 'executing',
@@ -185,10 +187,10 @@ class CourseDeleteServiceImpl extends BaseService implements CourseDeleteService
             return;
         }
         //当前系统已不存在这个job PushNotificationOneHourJob_lesson_taskId
-        $this->getSchedulerService()->deleteJobByName('PushNotificationOneHourJob_lesson_'.$task['id']);
-        $this->getSchedulerService()->deleteJobByName('LiveCourseStartNotifyJob_liveLesson_'.$task['id']);
-        $this->getSchedulerService()->deleteJobByName('SmsSendOneDayJob_task_'.$task['id']);
-        $this->getSchedulerService()->deleteJobByName('SmsSendOneHourJob_task_'.$task['id']);
+        $this->getSchedulerService()->deleteJobByName('PushNotificationOneHourJob_lesson_' . $task['id']);
+        $this->getSchedulerService()->deleteJobByName('LiveCourseStartNotifyJob_liveLesson_' . $task['id']);
+        $this->getSchedulerService()->deleteJobByName('SmsSendOneDayJob_task_' . $task['id']);
+        $this->getSchedulerService()->deleteJobByName('SmsSendOneHourJob_task_' . $task['id']);
     }
 
     protected function deleteCourseJob($courseId)
